fix: add gather_facts interface to connector
