+++ pkg/connector/local_connector.go
@@ -17,15 +17,21 @@ limitations under the License.
 package connector
 
 import (
+	"bytes"
 	"context"
+	"fmt"
 	"io"
 	"io/fs"
 	"os"
 	"path/filepath"
+	"runtime"
 
 	"k8s.io/klog/v2"
 	"k8s.io/utils/exec"
 )
+
+var _ Connector = &localConnector{}
+var _ GatherFacts = &localConnector{}
 
 type localConnector struct {
 	Cmd exec.Interface
@@ -68,4 +74,53 @@ func (c *localConnector) FetchFile(ctx context.Context, src string, dst io.Write
 func (c *localConnector) ExecuteCommand(ctx context.Context, cmd string) ([]byte, error) {
 	klog.V(4).InfoS("exec local command", "cmd", cmd)
 	return c.Cmd.CommandContext(ctx, "/bin/sh", "-c", cmd).CombinedOutput()
+}
+
+func (c *localConnector) Info(ctx context.Context) (map[string]any, error) {
+	switch runtime.GOOS {
+	case "linux":
+		// os information
+		osVars := make(map[string]any)
+		var osRelease bytes.Buffer
+		if err := c.FetchFile(ctx, "/etc/os-release", &osRelease); err != nil {
+			return nil, fmt.Errorf("failed to fetch os-release: %w", err)
+		}
+		osVars["release"] = convertBytesToMap(osRelease.Bytes(), "=")
+		kernel, err := c.ExecuteCommand(ctx, "uname -r")
+		if err != nil {
+			return nil, fmt.Errorf("get kernel version error: %w", err)
+		}
+		osVars["kernel_version"] = string(bytes.TrimSuffix(kernel, []byte("\n")))
+		hn, err := c.ExecuteCommand(ctx, "hostname")
+		if err != nil {
+			return nil, fmt.Errorf("get hostname error: %w", err)
+		}
+		osVars["hostname"] = string(bytes.TrimSuffix(hn, []byte("\n")))
+		arch, err := c.ExecuteCommand(ctx, "arch")
+		if err != nil {
+			return nil, fmt.Errorf("get arch error: %w", err)
+		}
+		osVars["architecture"] = string(bytes.TrimSuffix(arch, []byte("\n")))
+
+		// process information
+		procVars := make(map[string]any)
+		var cpu bytes.Buffer
+		if err := c.FetchFile(ctx, "/proc/cpuinfo", &cpu); err != nil {
+			return nil, fmt.Errorf("get cpuinfo error: %w", err)
+		}
+		procVars["cpuInfo"] = convertBytesToSlice(cpu.Bytes(), ":")
+		var mem bytes.Buffer
+		if err := c.FetchFile(ctx, "/proc/meminfo", &mem); err != nil {
+			return nil, fmt.Errorf("get meminfo error: %w", err)
+		}
+		procVars["memInfo"] = convertBytesToMap(mem.Bytes(), ":")
+
+		return map[string]any{
+			"os":      osVars,
+			"process": procVars,
+		}, nil
+	default:
+		klog.V(4).ErrorS(nil, "Unsupported platform", "platform", runtime.GOOS)
+		return nil, nil
+	}
 }
